+++ packages/nimble-components/src/mapping/base/types.ts
@@ -1,0 +1,1 @@
+export type MappingKey = string | number | boolean;