+++ packages/nimble-components/src/table-column/anchor/tests/table-column-anchor-matrix.stories.ts
@@ -16,7 +16,7 @@ import {
 } from '../../../theme-provider/design-tokens';
 
 const metadata: Meta = {
-    title: 'Tests/Table Column Types',
+    title: 'Tests/Table Column - Anchor',
     parameters: {
         ...sharedMatrixParameters()
     }