+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Wed, 02 Aug 2023 19:10:45 GMT and should not be manually modified.
+This log was last generated on Mon, 07 Aug 2023 22:54:53 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 17.0.10
+
+Mon, 07 Aug 2023 22:54:53 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.1.3
+
+## 17.0.9
+
+Thu, 03 Aug 2023 22:50:39 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.1.2
+
+## 17.0.8
+
+Thu, 03 Aug 2023 16:47:18 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v20.1.1
 
 ## 17.0.7